ceylonroots/admin-panel/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
bcrypt==4.2.0
